clean a bit notebook + move tests
